server/requirements.txt
FLASK
beautifulsoup4
langchain_community
langchain_openai
python-dotenv
youtube-transcript-api
chromadb
langchain-chroma
duckduckgo-search